Release v2.3

+++ src/authorization/apikeyauthorizer/apikeyauthorizer.decorator.ts
@@ -1,10 +1,16 @@
 import { createParamDecorator, ExecutionContext, ForbiddenException } from "@nestjs/common"
 import { SubscriptionModel } from "src/api/subscription/entities/subscription.entity"
+import { WorkspaceModel } from "src/api/workspace/entities/workspace.entity"
 import { apiPricing } from "src/config/subscriptionConfig"
 import { statusMessages } from "src/constants/statusMessages"
 
+export interface ApiKeyAuthorizerReturnType {
+  userId: string,
+  workspaceId: string
+}
+
 export const ApiKeyAuthorizer = createParamDecorator(
-  async (data: unknown, ctx: ExecutionContext) => {
+  async (data: unknown, ctx: ExecutionContext): Promise<ApiKeyAuthorizerReturnType> => {
     const request = ctx.switchToHttp().getRequest()
     const apiKey = request.headers["x-api-key"]
     const requestedResource = String(request.originalUrl).split("/")[2]
@@ -18,26 +24,36 @@ export const ApiKeyAuthorizer = createParamDecorator(
         const subscription = await SubscriptionModel.findOne({ apiKey })
 
         if (subscription) {
-          const currentDate = new Date()
-          const expiryDate = subscription.expiresAt
+          const workspaceId = subscription.workspaceId.toString()
+          const workspace = await WorkspaceModel.findById(workspaceId)
 
-          if (currentDate > expiryDate) {
-            await SubscriptionModel.findOneAndDelete({ apiKey })
-            throw new ForbiddenException(statusMessages.apiKeyExpired)
-          }
+          if (workspace) {
+            const userId = workspace.ownerId.toString()
+            const currentDate = new Date()
+            const expiryDate = subscription.expiresAt
 
-          else {
-            const creditRequiredForCurrentRequest = apiPricing[`${requestedResource}`]
-
-            if (creditRequiredForCurrentRequest > subscription.remainingCredits) {
-              throw new ForbiddenException(statusMessages.apiKeyLimitReached)
+            if (currentDate > expiryDate) {
+              await SubscriptionModel.findOneAndDelete({ apiKey })
+              throw new ForbiddenException(statusMessages.apiKeyExpired)
             }
 
             else {
-              subscription.remainingCredits -= creditRequiredForCurrentRequest
-              await subscription.save()
-              return subscription.owner.toString()
+              const creditRequiredForCurrentRequest = apiPricing[`${requestedResource}`]
+
+              if (creditRequiredForCurrentRequest > subscription.remainingCredits) {
+                throw new ForbiddenException(statusMessages.apiKeyLimitReached)
+              }
+
+              else {
+                subscription.remainingCredits -= creditRequiredForCurrentRequest
+                await subscription.save()
+                return { userId, workspaceId }
+              }
             }
+          }
+
+          else {
+            throw new ForbiddenException(statusMessages.invalidApiKey)
           }
         }
